12pm

+++ server/controllers/BlogController.js
@@ -45,7 +45,7 @@ export default class BlogController {
 
   async edit(req, res, next) {
     try {
-      let data = await _blogService.findOneAndUpdate({ _id: req.params.id, }, req.body, { new: true })
+      let data = await _blogService.findOneAndUpdate({ _id: req.params.id, authorId: req.session.uid }, req.body, { new: true })
       if (data) {
         return res.send(data)
       }
@@ -57,7 +57,10 @@ export default class BlogController {
 
   async delete(req, res, next) {
     try {
-      await _blogService.findOneAndRemove({ _id: req.params.id })
+      let data = await _blogService.findOneAndRemove({ _id: req.params.id, authorId: req.session.uid })
+      if (!data) {
+        throw new Error("invalid id")
+      }
       res.send("deleted blog")
     } catch (error) { next(error) }
